requirements.txt
setuptools~=57.0.0
argparse~=1.4.0
python-dotenv~=0.19.0